requirements.txt
astroid==2.2.5
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-admin==1.3.2
django-excel-response2==2.0.8
django-six==1.0.4
firebase-admin==2.17.0
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-core==1.0.2
google-cloud-firestore==1.3.0
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpcio==1.22.0
httplib2==0.13.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
msgpack==0.6.1
oauthlib==3.0.2
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pylint==2.3.1
python-firebase==1.2
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
screen==1.0.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
xlwt==1.3.0
